Moving assertHelpers under test/lib
