+++ test/specs/rules/disallow-comma-before-line-break.js
@@ -1,6 +1,6 @@
 var Checker = require('../../../lib/checker');
 var assert = require('assert');
-var reportAndFix = require('../../assertHelpers').reportAndFix;
+var reportAndFix = require('../../lib/assertHelpers').reportAndFix;
 
 describe('rules/disallow-comma-before-line-break', function() {
     var rules = { disallowCommaBeforeLineBreak: true };
