+++ test/specs/rules/disallow-keywords-on-new-line.js
@@ -1,6 +1,6 @@
 var Checker = require('../../../lib/checker');
 var assert = require('assert');
-var reportAndFix = require('../../assertHelpers').reportAndFix;
+var reportAndFix = require('../../lib/assertHelpers').reportAndFix;
 
 describe('rules/disallow-keywords-on-new-line', function() {
     var checker;
