+++ test/specs/rules/disallow-padding-newlines-after-use-strict.js
@@ -1,6 +1,6 @@
 var Checker = require('../../../lib/checker');
 var assert = require('assert');
-var reportAndFix = require('../../assertHelpers').reportAndFix;
+var reportAndFix = require('../../lib/assertHelpers').reportAndFix;
 
 describe('rules/disallow-padding-newlines-after-use-strict', function() {
     var rules = { disallowPaddingNewLinesAfterUseStrict: true };
